Update mask_attn_weights to work with mixed precision training
Since we are multiplying by 1e10 originally to mask attention weights, when training in fp16, there is a gradient overflow. I've changed this to 1e4 to keep it within the range of fp16, and now it works in mixed precision. I've also tested with different seeds and the unconditional generation is unaffected. Please let me know if there anything else I need to do to complete this pull request. Thanks for releasing this model!